chore: Merge all DatePicker into a unique component
