--- packages/components/src/DateTimePickers/DatePicker/CalendarBody/CalendarBody.scss
@@ -1,15 +1,0 @@
-
-.container {
-    display: flex;
-    flex-direction: column;
-
-    .week {
-        display: flex;
-
-        .day {
-            padding: 10px;
-            width: 2.5em;
-            text-align: center;
-        }
-    }
-}
