--- packages/components/src/DateTimePickers/DatePicker/CalendarBody/CalendarBody.test.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import { shallow } from 'enzyme';
-
-import CalendarBody from './CalendarBody.component';
-
-describe('CalendarBody', () => {
-	it('should render a CalendarBody', () => {
-		// when
-		const wrapper = shallow(<CalendarBody />);
-
-		// then
-		expect(wrapper.getElement()).toMatchSnapshot();
-	});
-});
