--- packages/components/src/DateTimePickers/DatePicker/CalendarBody/index.js
@@ -1,3 +1,0 @@
-import CalendarBody from './CalendarBody.component';
-
-export default CalendarBody;
