--- packages/components/src/DateTimePickers/DatePicker/CalendarHeader/CalendarHeader.component.js
@@ -1,19 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import theme from './CalendarHeader.scss';
-
-function CalendarHeader(props) {
-	const dayNames = ['L', 'M', 'M', 'J', 'V', 'S', 'D'];
-	return (
-		<div className={theme.container}>
-			{dayNames.map((dayName, i) =>
-				<abbr className={theme['day-of-week']} key={i}>{dayName}</abbr>
-			)}
-		</div>
-	);
-}
-
-CalendarHeader.propTypes = {
-};
-
-export default CalendarHeader;
