--- packages/components/src/DateTimePickers/DatePicker/CalendarHeader/CalendarHeader.scss
@@ -1,10 +1,0 @@
-
-.container {
-    display: flex;
-
-    .day-of-week {
-        padding: 10px;
-        width: 2.5em;
-        text-align: center;
-    }
-}
