--- packages/components/src/DateTimePickers/DatePicker/CalendarHeader/index.js
@@ -1,3 +1,0 @@
-import CalendarHeader from './CalendarHeader.component';
-
-export default CalendarHeader;
